Move variable results helper functions
