+++ src/components/search/results/variable-results/index.js
@@ -1,3 +1,4 @@
 export * from './variable-results'
+export * from './variable-results-helper'
 export * from './variable-table-by-study'
 export * from './variables-histogram'
